techDebt(bulkUploadWorkersRefactor): Move validateDuplicateWorkerID tests into separate spec file
