+++ server/test/unit/routes/establishments/bulkUpload.spec.js
@@ -18,9 +18,6 @@ const {
   validateDuplicateLocations,
 } = require('../../../../routes/establishments/bulkUpload/validate/validateDuplicateLocations');
 const {
-  validateDuplicateWorkerID,
-} = require('../../../../routes/establishments/bulkUpload/validate/workers/validateDuplicateWorkerID');
-const {
   validatePartTimeSalary,
 } = require('../../../../routes/establishments/bulkUpload/validate/workers/validatePartTimeSalary');
 
@@ -240,76 +237,6 @@ describe('/server/routes/establishment/bulkUpload.js', () => {
     });
   });
 
-  describe('validateDuplicateWorkerID()', () => {
-    it('errors when CHGUNIQUEWRKID is not unique', async () => {
-      const csvWorkerSchemaErrors = [];
-      const allWorkersByKey = {};
-      const myAPIWorkers = [];
-      const myWorkers = [
-        buildWorkerCSV({
-          overrides: {
-            LOCALESTID: 'foo',
-            UNIQUEWORKERID: 'Worker 1',
-          },
-        }),
-        buildWorkerCSV({
-          overrides: {
-            LOCALESTID: 'foo',
-            UNIQUEWORKERID: 'Worker 2',
-            CHGUNIQUEWRKID: 'Worker 1',
-          },
-        }),
-      ].map((currentLine, currentLineNumber) => {
-        const worker = new WorkerCsvValidator.Worker(currentLine, currentLineNumber, []);
-
-        worker.validate();
-
-        return worker;
-      });
-
-      const allKeys = myWorkers.map((worker) => (worker.local + worker.uniqueWorker).replace(/\s/g, ''));
-
-      myWorkers.forEach((thisWorker) => {
-        // uniquness for a worker is across both the establishment and the worker
-        const keyNoWhitespace = (thisWorker.local + thisWorker.uniqueWorker).replace(/\s/g, '');
-        const changeKeyNoWhitespace = thisWorker.changeUniqueWorker
-          ? (thisWorker.local + thisWorker.changeUniqueWorker).replace(/\s/g, '')
-          : null;
-
-        if (
-          validateDuplicateWorkerID(
-            myWorkers[1],
-            allKeys,
-            changeKeyNoWhitespace,
-            keyNoWhitespace,
-            allWorkersByKey,
-            myAPIWorkers,
-            csvWorkerSchemaErrors,
-          )
-        ) {
-          allWorkersByKey[keyNoWhitespace] = thisWorker.lineNumber;
-
-          // to prevent subsequent Worker duplicates, add also the change worker id if CHGUNIQUEWORKERID is given
-          if (changeKeyNoWhitespace) {
-            allWorkersByKey[changeKeyNoWhitespace] = thisWorker.lineNumber;
-          }
-        }
-      });
-
-      expect(csvWorkerSchemaErrors.length).equals(1);
-      expect(csvWorkerSchemaErrors[0]).to.eql({
-        origin: 'Workers',
-        lineNumber: 1,
-        errCode: 998,
-        errType: 'DUPLICATE_ERROR',
-        error: 'CHGUNIQUEWRKID Worker 1 is not unique',
-        name: 'foo',
-        source: 'Worker 2',
-        worker: 'Worker 2',
-        column: 'CHGUNIQUEWRKID',
-      });
-    });
-  });
   describe('validatePartTimeSalary()', () => {
     // FTE / PTE : Full / Part Time Employee . FTE > 36 hours a week, PTE < 37
     it('errors when one worker has the same salary as a FTE but works PTE', async () => {
